Keep WrapperOpaque in own file
